+++ src/DGC.Sample.Domain/Specifications/Orders/OrderFilterSpec.cs
@@ -6,7 +6,7 @@ namespace DGC.Sample.Domain.Specifications.Orders;
 public sealed class OrderFilterSpec : Specification<Order>
 {
     public OrderFilterSpec(OrderStatus? status, string? customerNameSearch)
-        : base(o => (!status.HasValue || o.Status == status.Value) &&
+        : base(o => (!status.HasValue || o.Status == (int)status.Value) &&
                     (string.IsNullOrWhiteSpace(customerNameSearch) || o.CustomerName.Contains(customerNameSearch)))
     {
         ApplyOrderByDescending(o => o.OrderDateUtc);
